# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Order:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      order_number:
        type: string
      payment_method:
        $ref: '#/definitions/domain.PaymentMethod'
      payment_status:
        $ref: '#/definitions/PaymentStatus'
      status:
        $ref: '#/definitions/domain.OrderStatus'
      total_price:
        type: number
      updated_at:
        type: string
      user:
        $ref: '#/definitions/User'
      user_id:
        type: string
    type: object
  PaymentStatus:
    enum:
    - PENDING
    - PAID
    - REFUNDED
    - FAILED
    type: string
    x-enum-varnames:
    - PaymentPending
    - PaymentPaid
    - PaymentRefunded
    - PaymentFailed
...
  User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login:
        type: string
      last_name:
        type: string
      mobile:
        type: string
      mobile_verified:
        type: boolean
      role:
        $ref: '#/definitions/UserRole'
      updated_at:
        type: string
    type: object
...
  UserRole:
    enum:
    - USER
    - ADMIN
    type: string
    x-enum-varnames:
    - UserRoleUser
    - UserRoleAdmin
...
  domain.CartItemDetail:
    properties:
      author_name:
        type: string
      book_id:
        type: string
      count:
        type: integer
      image_url:
        type: string
      line_total:
        type: number
      name:
        type: string
      unit_price:
        type: number
    type: object
  domain.CartItemInput:
    properties:
      book_id:
        type: string
      count:
        minimum: 1
        type: integer
    required:
    - book_id
    - count
    type: object
  domain.CartView:
    properties:
      cart_id:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.CartItemDetail'
        type: array
      subtotal:
        type: number
      total_items:
        type: integer
      user_id:
        type: string
    type: object
...
  domain.CheckoutInput:
    properties:
      payment_method:
        $ref: '#/definitions/domain.PaymentMethod'
    required:
    - payment_method
    type: object
  domain.OrderItemDetail:
    properties:
      book_id:
        type: string
      count:
        type: integer
      image_url:
        type: string
      line_total:
        type: number
      name:
        type: string
      unit_price:
        type: number
    type: object
  domain.OrderStatus:
    enum:
    - PENDING
    - CANCELLED
    - COMPLETED
    type: string
    x-enum-varnames:
    - OrderPending
    - OrderCancelled
    - OrderCompleted
  domain.OrderView:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.OrderItemDetail'
        type: array
      order:
        $ref: '#/definitions/Order'
    type: object
  domain.PaymentConfirmInput:
    properties:
      order_id:
        type: string
      success:
        type: boolean
    required:
    - order_id
    type: object
  domain.PaymentMethod:
    enum:
    - COD
    - CREDIT_CARD
    - DEBIT_CARD
    - NET_BANKING
    - UPI
    type: string
    x-enum-varnames:
    - PaymentCOD
    - PaymentCreditCard
    - PaymentDebitCard
    - PaymentNetBanking
    - PaymentUPI
host: localhost:8080
info:
  contact:
    email: kolimirakshitha@gmail.com
    name: Rakshitha
  description: Online Bookstore backend (BookNest)
  license:
    name: MIT
  termsOfService: http://booknest.com/terms/
  title: BookNest API
  version: "1.0"
paths:
  /admin/orders:
    get:
      description: Lists all orders (admin only)
      parameters:
      - description: Result limit
        in: query
        name: limit
        type: integer
      - description: Result offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.OrderView'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all orders
      tags:
      - Orders
...
  /cart:
    get:
      description: Fetches the authenticated user's cart with items and totals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CartView'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cart
      tags:
      - Cart
  /cart/clear:
    post:
      description: Removes all items from the authenticated user's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Clear cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Adds a book to the authenticated user's cart
      parameters:
      - description: Cart item input
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.CartItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CartView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Updates quantity for a cart item in the authenticated user's cart
      parameters:
      - description: Cart item input
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.CartItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CartView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update cart item
      tags:
      - Cart
  /cart/items/{book_id}:
    delete:
      description: Removes a book from the authenticated user's cart
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CartView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
  /orders:
    get:
      description: Lists orders for the authenticated user
      parameters:
      - description: Result limit
        in: query
        name: limit
        type: integer
      - description: Result offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.OrderView'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List my orders
      tags:
      - Orders
  /orders/checkout:
    post:
      consumes:
      - application/json
      description: Creates an order from the authenticated user's cart
      parameters:
      - description: Checkout input
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.CheckoutInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.OrderView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Checkout order
      tags:
      - Orders
  /orders/confirm:
    post:
      consumes:
      - application/json
      description: Confirms payment for an existing order
      parameters:
      - description: Payment confirmation input
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.PaymentConfirmInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.OrderView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Confirm payment
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"